chore: make interfaces consistent
