+++ packages/codemirror-graphql/src/__tests__/hint-test.js
@@ -40,7 +40,7 @@ function createEditorWithHint() {
       schema: TestSchema,
       closeOnUnfocus: false,
       completeSingle: false,
-      externalFragments: [],
+      externalFragments: 'fragment Example on Test { id }',
     },
   });
 }
@@ -767,6 +767,11 @@ describe('graphql-hint', () => {
         type: TestType,
         description: 'fragment Foo on Test',
       },
+      {
+        text: 'Example',
+        type: TestType,
+        description: 'fragment Example on Test',
+      },
     ];
     const expectedSuggestions = getExpectedSuggestions(list);
     expect(suggestions.list).to.deep.equal(expectedSuggestions);
@@ -782,6 +787,11 @@ describe('graphql-hint', () => {
         text: 'Foo',
         type: TestType,
         description: 'fragment Foo on Test',
+      },
+      {
+        text: 'Example',
+        type: TestType,
+        description: 'fragment Example on Test',
       },
     ];
     const expectedSuggestions = getExpectedSuggestions(list);
