+++ packages/codemirror-graphql/src/hint.js
@@ -13,7 +13,7 @@
 import CodeMirror from 'codemirror';
 import { getAutocompleteSuggestions } from 'graphql-language-service-interface';
 import { Position } from 'graphql-language-service-utils';
-
+import { getFragmentDefinitions } from './utils/getFragmentDefinitions';
 /**
  * Registers a "hint" helper for CodeMirror.
  *
@@ -52,7 +52,9 @@ CodeMirror.registerHelper('hint', 'graphql', (editor, options) => {
     editor.getValue(),
     position,
     token,
-    options.externalFragmentDefinitions || undefined,
+    Array.isArray(options.externalFragments)
+      ? options.externalFragments
+      : getFragmentDefinitions(options.externalFragments),
   );
 
   const results = {
