+++ packages/graphiql/src/components/QueryEditor.tsx
@@ -12,8 +12,6 @@ import {
   GraphQLSchema,
   GraphQLType,
   ValidationRule,
-  visit,
-  parse,
 } from 'graphql';
 import MD from 'markdown-it';
 import { normalizeWhitespace } from '../utility/normalizeWhitespace';
@@ -40,15 +38,6 @@ type QueryEditorProps = {
   externalFragments?: string | FragmentDefinitionNode[];
 };
 
-function gatherFragmentDefinitions(graphqlString: string) {
-  const definitions: FragmentDefinitionNode[] = [];
-  visit(parse(graphqlString), {
-    FragmentDefinition(node) {
-      definitions.push(node);
-    },
-  });
-  return definitions;
-}
 /**
  * QueryEditor
  *
@@ -101,27 +90,6 @@ export class QueryEditor extends React.Component<QueryEditorProps, {}>
     require('codemirror-graphql/jump');
     require('codemirror-graphql/mode');
 
-    const hintOptions: {
-      schema?: GraphQLSchema;
-      externalFragmentDefinitions?: FragmentDefinitionNode[];
-      closeOnUnfocus: boolean;
-      completeSingle: boolean;
-      container: any;
-    } = {
-      schema: this.props.schema,
-      closeOnUnfocus: false,
-      completeSingle: false,
-      container: this._node,
-    };
-
-    if (this.props.externalFragments) {
-      hintOptions.externalFragmentDefinitions = Array.isArray(
-        this.props?.externalFragments,
-      )
-        ? this.props.externalFragments
-        : gatherFragmentDefinitions(this.props.externalFragments);
-    }
-
     const editor: CM.Editor = (this.editor = CodeMirror(this._node, {
       value: this.props.value || '',
       lineNumbers: true,
@@ -139,8 +107,16 @@ export class QueryEditor extends React.Component<QueryEditorProps, {}>
       lint: {
         schema: this.props.schema,
         validationRules: this.props.validationRules ?? null,
+        // linting accepts string or FragmentDefinitionNode[]
+        externalFragments: this.props?.externalFragments,
       },
-      hintOptions,
+      hintOptions: {
+        schema: this.props.schema,
+        closeOnUnfocus: false,
+        completeSingle: false,
+        container: this._node,
+        externalFragments: this.props?.externalFragments,
+      },
       info: {
         schema: this.props.schema,
         renderDescription: (text: string) => md.render(text),
